Pygame-RPG 22 Consolidating Object and NPC classes
This PR focuses on simplifying the `Object` and `NPC` class by creating
a parent class for them. We also wish to simplify their manager classes
as well by creating a parent class for them (`NPCManager` and
`ObjectManager) as well.

The parent class for `NPC` and `Object` is `Interactable`. The reason
why `Interactable` was created was because `NPC` and `Object` have
nearly identical code with some minute differences. So to reduce the
redundancy, this abstract class was created.

`Interactable` inherits from the `Sprite` class and has the following
attributes, `Name`. `Sprite`, `aniStatus`, `genericKey`, `Scale`,
`Flipped`, `Context`, `Pos`, `Events`, `aniTracker`, `maxAniVal`,
`image` and `rect`. The only attribute that is new, is `genericKey`.

`genericKey`: is an event key that points to an event for when the
`Interactable` object is interacted with, but the list of events is
empty.

The `Interactable` class has the following functions,
`get_updated_data()`, `process_special_criteria()`,
`is_colliding()`, `process_collision()`, `super_get_max_ani_val()`,
`set_image_and_rect()`, `super_update()`, `add_event_keys()`,
`get_event_key()`.

`get_updated_data()`: This is a function that returns a dictionary of
values that were changed for the given `Interactable`. This function is
meant to be overwritten by the child classes.

`process_special_criteria()`: This function evaluates any special
criteria for the `aniTracker` attribute to be set to -1 i.e. stuck.
This function is meant to be overwritten by the child classes.

`is_colliding()`: This function returns whether the `Interactable` is
colliding with the knight object or not. This function is meant to be
overwritten by the child classes.

`process_collision()`: This function runs when an `Interactable` was
interacted with and determines what the `Interactable` does (if it does
anything). This function is meant to be overwritten by the child
classes.

`super_get_max_ani_val()`: This function is just a fancier way of
calling `get_max_ani_val()`.

`set_image_and_rect()`: This function sets the `image` and `rect` attributes
of an `Interactable` class. This class now takes in a name to construct
the file path however.

`super_update()`: A mask that actually deals with updating the sprite
data. takes in a string name for `set_image_and_rect()` and a boolean,
force, to see if the update should be forced.

`add_event_keys()`: This function takes in a list of events and adds
them to the event list of the `Interactable`

`get_event_key()`: This function returns the first item in the `Events`
list (if there is one) or returns the `genericKey` attribute.

As a result of this, these changes, `NPC` has the `Flipped` attribute
and the `Dialogue` attribute is renamed to `Events` since `Events` is
inherited by the parent class. The generic key for all NPCs (at the
moment) is "genericDialogue". `Object` had similar changes to it since
both classes now inherit from `Interactable`.

For `InteractableManager` it has the attributes, `interactableDict`,
`interactableGroup` and `knight`. These mirror the attributes of it's
child classes.

It has the functions
`update_interactable()`, `save_and_empty()`, `get_colliding()` and
`get_interaction_event()`.

`update_interactable()`: Just a renamed version of `update_NPC()` or
`update_object()`.

`save_and_empty()`: Same as before but now calls on the sprite's
`get_updated_data()` function to get the dictionary it needs to update.

`get_colliding()`: Runs through every sprite in a group and calls the
`is_colliding()` function, if the function returns true, the sprite is
returned.

`get_interaction_event()`: If an interaction is confirmed, the sprite's
`process_collision()` function is called.

Because of these changes, the `NPCManager` and `ObjectManager` class
only have the functions related to the class they're managing such as
`create_object()`/`create_NPC()`, `get_objects()`/`get_NPCs()`.

Other changes:
- ObjectManager is now integrated with Save_Manager
- schema changes in JSON
- updates to the scripts
- updates to the test files and test cases
- Updated save files
- renamed the attributes of `ObjectManager` and `NPCManager`
- `NPC` class now supports flipped sprites