Pygame-RPG 23 Revamp Event Queue for Interactables
The interactions with the `Interactable` classes, aren't fully
implemented because they don't align with how they're defined in the
`Events.json` file. In other words, all events are treated as one time
events and are tossed off the queue.

To fix this, there needs to be a check in place for if an event is
repeatable. To do this, the `get_cut_event_info()` function was created
in the `Utils` package. This function returns a dictionary of the event
keys mapped to another dictionary with 2 keys, "Priority" and
"Repeatable".

Now with this change, each `Interactable` Object gets a reference to
this dictionary as the `simpleEventDict` attribute.
Using this, `get_event_key()` first checks if the first
event key in the queue is of a repeatable event. If it is, only the key
is returned, the size of the queue does not change (i.e. we don't call
`pop()` to dequeue the first element). Otherwise, we use `pop()` to
dequeue the latest event key.

This change, while helpful, creates a problem. An event key could be
stuck behind an repeatable event key. This is a problem since the game
should update NPC dialogue after a major event occurs.

To amend this, a priority key was added to the event schema.
This priority system only matters for the dialogue events, any other
type of event has a default priority of -1, the highest possible.
Meanwhile, repeatable dialogue events have the lowest possible priority
at 100 (remember, `sort()` does things in ascending order. For a queue
that means the higher the priority number, the farther it is from
the start).

The priority is used in the `sort_events()` function in the
`Interactable` class. As a result of this, I needed to make the
`Interactable_test.py` class to test this. `sort_events()` is called
when an interactable class `NPC` or `Object` is initialized and when the
`add_events()` function is called.

Other changes:
- Changes in the test file of the `Interactable` classes
- Changes to the JSON schema
- Updated the Scripts.py file to create new schema for events
- Updated save files
- added the missing text file from event_text/