Pygame-RPG 23.1 Consolodating Entity
This PR is just to move all of the animation work from the `Knight`
class over to the `Entity` class. This is so that eventually, `Enemy`
inherits the functions needed to be animated. The functions moved from
`Knight.py` to `Entity.py` are the following, `set_image_and_rect()` and
`update()`

This PR also adds the attributes for how animation works over from `Knight`
to `Entity`. These attributes are the following, `x`, `y`, `aniStatus`,
`aniTracker`, `maxAniVal`, `image`, `rect` and `flipped`.

Other changes:
- Updated the JSON schemas of enemy objects
- slight optimizations to `Rpg2.py` (the `handle_basic_input()` function)
- Added tests for Entities
- `update()` also updates the rectangles position
- Changed the name of the script that adds enemy schemas to "add-enemy"
  instead of "create-enemy" on both `script` and `Scripts.py`