Pygame-RPG 24 Creating Animation Manager
This PR is for creating the `AnimationManager`. This class is
responsible for animating the actions of the Entity classes in battle.

The `AnimationManager` class has the following attributes:

`screen`: screen object

`active`: A boolean that indicates whether the manager is active or not

`spriteGroup`: A pygame sprite group

`actor`: The Sprite object that is performing the actions

`targets`: A list of Sprite objects that will be affected by the actions
of the actor

`others`: The rest of the Sprite objects in a battle that are not
affected by the actions of the actor

`actorStartingPos`: A tuple representing the x and y coordinate of the
actor before moving. This is so that the actor can return there.

`actionQueue`: A list of actions that the manager needs to perform
before becoming inactive

`frameCount`: A number representing how many frames have passed

`maxFrameCount`: A number representing the maximum amount of frames that
can pass for a given action.

`action`: The action being performed at the moment.

`slopeEquation`: A dictionary containing the information needed for the
slope equation. This includes "m" (slope), "b"(y intercept), "Starting
Point", "x_dsitance" (the distance between the starting and end points).

`x_increment`: How much to add to the x value of the actor for every
frame

This class has the following functions, `set_slope_equation()`,
`use_slope_equation()`, `confirm_valid_animation_status()`,
`fill_sprite_group()`, `change_targets_animation()`,
`load_action_queue()`, `apply_new_action()`, `reset_action()` and
`process_action()`.

`set_slope_equation()`: This function takes two tuples of x and y
coordinates and uses it to form the slope line equation and fills out
the necessary information in the `slopeEquation` dictionary.

`use_slope_equation()`: This function uses the `slopeEquation`
dictionary and the `frameCount` attribute to set the new x and y values
of the actor.

`confirm_valid_animation_status()`: Takes in the string file path and an
animation status string. It then uses Regex to see if there is a match
in the png files of the folder. If there is a match, the function
returns true, otherwise, it returns false.

`fillSpriteGroup()`: This function takes in a list of sprites and adds
them to the `spriteGroup` attribute.

`change_targets_animations()`: This function takes in a list of sprites
along with a string animation status that the user wants to apply to
every sprite. If the sprite isn't already doing the animation or isn't
dead, their animation information is reset and the `aniStatus` attribute
is changed.

`load_action_queue()`: Takes in a move object, Entity object, a list of
sprites that are affected by the animation and a list of other sprites.
This function then uses those attributes to create the action
dictionaries (dictionaries that describe what an action does) and add it
to the `actionQueue`. This function also sets all the other attributes
up as well.

`apply_new_action`: This function applies the new action so that the
animation can be properly processed.

`reset_action()`: Reset the manager

`process_action()`: Draws the group of sprites while processing the
animation

Other Changes:
- Updates to the `add_enemy()` function in `BattleManager`
- Update the `Move` class to have the new attributes from the dictionary
- Updated JSON schemas
- changed the name of the command for scripts from "create-move" to
  "add-move"
- Updated Scripts