Pygame-RPG 25 Integrating Animation Manager
This PR focuses on integrating the `AnimationManager` class into the
`Rpg2.py` file. First however, there were other changes in other files
that needed to be made.

`Entity`:
This class received 2 new attributes, `altName` and `Scale`, `altName`
is a string that's an alternative to the `Name` attribute. The reason
for this division is that `Name` is used for animation so it can't be
changed but sometimes we do want to call objects by a different name to
differentiate them (for example in combat "Goblin" and "Goblin2").
Scale, is a tuple that represents the size and width of the sprite. This
works the same as the `Scale` attribute found in `Interactable`.

`Enemy`:
This class now takes in the `Scale` attribute from the JSON file

`Knight`:
now has a set `Scale` attribute on init

`EntityFactory`: When using the `create_entity()` functions, the
function no longer changes the `Name` attribute but the `altName`
attribute.

`AnimationManager`:
This class received many changes for the sake of the integration. Below
are the notable changes.

`use_slope_equation()`: The condition for checking if the equation is
proper is changed to instead see if a starting point was present since
the `m`, `b` keys can be 0. `x_distance` also doesn't need to be checked
since it is set when the `startingPoint` key is set.

`load_action_queue()`: When this function is called, it now sects the
`active` attribute to True. The point the actor goes to for a
"Physical" attack is now changed to be dependent on an the inequality
based on the x coordinate of the actor and the target. If the x
coordinate of the actor is greater than the target. The way the "ranged"
attack point is also slightly different.

`partial_reset_action()`: A new function that resets some attributes
while leaving others. This is used when an action has been completed.

`reset_action()`:
A null check for the actor is added for when we attempt to call
`change_targets_animation()` during this function. This change is just
so that tests don't fail in `BattleManager`.

`process_action()`: The manager now resets everything when the queue
runs out of actions. The `use_slope_equation()` is null checked.
Finally, the `partial_reset_action()` function is called when the
`frameCount` reaches the `maxFrameCount`

`Algos.py`: This file received a new function called `get_others()`:
This function takes in a Sprite actor, list of Sprites and a list of all
Sprites. It then bassically creates a list of sprites that isn't the
actor and isn't in the list of selected targets.

`BattleManager`:
`BattleManager` now takes in a reference to the animationManager object.
and there were some changes made to the functions in the class.

`add_enemy()`: The function now centers the entity on the changed x and
y coordinates.

`do_one_turn()`: This function now calls the `load_animation_queue()`
function from `AnimationManager` when a proper move has been made. Also,
the `get_others()` function is used to get the uninvolved Entities.

The final change for the `BattleManager` class is that all returnable
strings now use `altName` instead of `Name`.

`Rpg2.py`: To support the `AnimationManager` class, the mock battle code
has been changed to make more sense. Finally, a condition to check
whether the `AnimationManager` is in use or the `BattleManager` is in
use. When the `AnimationManager` is in use, the `DialogueManager` is
also in use, meaning the dialogue and the animations run at the same
time. When the animation is done, the `BattleManager` part of the code
is run and when a move is made, the `AnimationManager` gets a turn.

Other Changes:
- Updates to JSON Schema
- Updates to testing files
- Slight change to how the cursor is displayed in `UIManager` for the
  "Select Targets" screen