Pygame-RPG 25.1 Housekeeping and bugfixing
This change is about fixing the animations when an item is used and
addressing a bug with when the item kills all enemies after the
integration.

Firstly, the animations for when an item is used cannot just be a static
animation. The reason for this is simple, some items should have
differing effects on what animations are played. For example, using a
damaging item should make the target go into a "Hurt" animation while
using a healing item shouldn't. Currently however, our
`AnimationManager` cannot handle this because the `load_action_queue`
function get's the animation information from a `Move` object.

To address this, the `load_action_queue()` no longer takes in a `Move`
object but a structured dictionary containing all of the key-pairs that
are needed to set up an animation.

Because of this, the `Move` class received a new function,
`get_action_details()`, which returns a structured dictionary based off
of the given `Move` object.

For items, however, the `Item_Effects.json` file needed to have a new
key called "Type" which is an indicator used in `AnimationManager`.
Also, `ItemManager` has a new function, `get_action_details()`. Unlike
the same function in the `Move` class, this function takes in a string
name and then returns the structured dictionary.

Secondly, the afformentioned bug was solved by simply adding a condition
for when the animation manager was inactive to check the battle's state
using the `determine_battle_state()` function from `BattleManager`.

Other Changes:
- Updated .idea files since I'm on a new computer
- Updated the way the targets were handled for animation manager in
  Rpg2.py
- Updated tests
- Updated scripts