Pygame-RPG 26 Fixing Animation Height
This PR is meant to address an oversight in the code. Previously, the
line followed by the actor to their target, another sprite object, was
drawn using the center of the actor's rectangle as a starting point and
the target's center as the end point.

The problem with this is that the sprites all have unique shapes and
sizes which causes the center of their rectangles to have unique quirks
as well. As a result, the animations did not come out as intended as the
sprites were not in line with each other when delivering their attacks.

To solve this, instead of basing the movement of the entities off of the
center of the rectangles, instead, we base it off the middle bottom of
their rectangles. By doing this, they are more in line with the original
vision. As a result of doing this, the default y value for the `Knight`
object was changed from 590, to 740.

As of the moment, I have no plans on implementing the same change to the
other `Sprite` (`Object` and `NPC`) mostly because I don't need too. In
the future, to streamline things I might. Also, due to this change,
`BattleManager` will need some changes. While the animations play out
fine, the cursor location as well as the location of the Entities are
off.

Other Changes:
- Updated save files
- Updated test files