Pygame-RPG 27 Overhaul BattleManager
This PR is focused on integrating the latest changes from the entity
classes over to `BattleManager` class. Early on in the design,
`BattleManager` kept track of the positions AND the entities tied to
them using a tuple. Now, since `Entity` classes track their own
position, this is unnecessary.

As a result, the `enemies` list now only contains the `Enemy` objects.
Because of this change, the functions using the coordinate part of the
previous tuple list now use the center of the `Enemy` object's rectangle
instead.

Furthermore, now that the `x` and `y` attributes of `Entity` classes are
correlating to the middle bottom part of their rectangle, the position
of the enemy and `Knight` objects are off. To fix this, `BattleManager`
centers the sprites around their point before setting the `x` and `y`
attribute.

Also, the `get_knight_pos()` function was created. It's just a simple
getter function that returns the center of the Object's rectangle.

Other Changes:
- Renamed `hero` attribute to `knight` in `BattleManager`
- Updated tests
- Updated where the cursor is drawn in `UIManager`
- Properly centered the `Knight` object in `Rpg2.py` when a battle
  starts