Pygame-RPG 27.1 Housekeeping and bugfixing
This PR focuses on addressing some bugs in the `AnimationManager` class.
It has been noticed that if the player enters another battle after
completing a prior one, the sprites for the old opponents are still
visible. This is because the sprite Group was improperly cleared.
Instead of using the `empty()` method, `remove()` was used instead in
the `reset()` function. By now using `empty()`, this bug no longer occurs.

The second bug that this PR addresses is a slight visual bug regarding
animations. Essentially, because update only works on 10 frame
intervals, this results in the last played animation to linger on screen
before it is changed. To fix this, we now call the `update()`
function for the `Entity` object and force the update to happen to
forcefully reset the sprite in use. This change applies to the
`change_targets_animation()` function.

This PR also has some optimizations for the `AnimationManager` class.
The `load_action_queue()` function now preloads the first action in the
queue (since when this function is called there is at minimum, 1 item in
the queue). The `process_action()` function has been streamlined so that
it checks if there's a new action available when the previous action has
finished. If there are more actions, it loads it and applies the action,
otherwise, it resets the manager. As a result of this, `BattleManager`
no longer resets the `ActionManager` in the `determine_battle_state()`
function since `ActionManager` would have already reset itself by the
time the function is run.

Other Changes:
- `reset_action()` --> `reset()` in `AnimationManager`
- `partial_reset_action()` --> `partial_reset()` in `AnimationManager`
- Added and modified tests