Pygame-RPG 27.2 Housekeeping and bugfixing
This PR is meant to address a simple oversight. Forced updates (calling
`update()` and giving it the True flag) is meant to correct the
animation of the `Sprite` object, nothing else. However, forced updates
also update the `aniTracker` attribute of the `Sprite` object which was
unintended. To fix this, I simply added a new condition for the
`aniTracker` to be incremented.

Other Changes:
- Updated tests