Pygame-RPG 28.1 Housekeeping and bugfixing
The following changes are:

`fix_entity_stats()` is called in `apply_effects()` instead of at the
end of `do_one_turn()` in `BattleManager`. The reason for this is
because the latter leads to a visible glitch where an Entity can "Over Heal"
(have more health than they should be allowed) before it get's fixed.

In `Rpg2.py`, `DialogueManager` now adds the list of strings from
`BattleManager` to the `backlog` list instead of the `dialogue` list. This is
because I believe that it would be better to communicate this
information visually rather than through text. However, a battle log is
going to be implemented eventually so this functionality isn't useless,
it's just put aside for now.

I also removed some comments from `DialogueManager`.