Pygame-RPG 28.2 Creating Animation Viewer
As the scope of this project grows, it's getting more difficult to
verify if the animations as well as the size of the sprites are correct.
For example, a dragon shouldn't be smaller than the knight. This PR
fixes this by creating the `Animation_Viewer.py` file. This file is used
to iterate through all of the sprites in the `Entity_Sprites` directory
and giving the user the option to choose a name.

Once the name is selected, the sprite is shown on screen and the user
can choose which animation they'd like the sprite to perform (Default is
"Idle").

This file is created using recycled code and some dumbed down versions
of both UIManager and UIHandler.

Other Changes:
- Edited the JSON files of some enemies to point the `sprite` attribute
  to the correct folders
- Edited the `Scripts.py` file so that upon creation, an enemy's
  `sprite` attribute will point to the correct folder.