Pygame-RPG 29 Displaying Damage During Battles
This PR is about creating a system to ensure that the changes in health for any
given entity takes is shown on screen to the player. Earlier on, this
was done through text and `DialogueManager`, this however, is cumbersome
and annoying to parse through (I don't like spamming enter every time I
make a move).

The first step to making this a reality is to obtain the differences in
health for each entity (changes in health being currentHealth -
initialHealth for the given turn). This is done through two methods
added in `BattleManager`, `get_initial_health()` and
`get_health_difference`.

`get_initial_health()`: This function returns a dictionary where the
`altName` attribute of the entity is mapped to a dictionary with the `Hp`
key that is mapped to their current health (i.e the `Hp` attribute).

`get_health_difference()`: This function takes in a dictionary(the same
one returned from `get_initial_health()`) and loops through the list of
all entities stored in `BattleManager` (i.e `knight` + `enemies` list).
It then compares their current health to the health stored in the
inputted dictionary. If there is a difference (i.e `currentHealth` !=
`initialHealth`) then an entry is made to a dictionary where entity's
`altName` attribute is mapped to a dictionary containing 2 keys, "Hp"
and "Pos". "Hp" is mapped to the difference between the current health
and the initial health of the Entity and the "Pos" key is mapped to
`entity.rect.midtop`.

The two above functions are used in the `do_one_turn()` function.
`get_health_info()` is used at the start of the function and the
information is stored. `get_health_difference()` is used at the end in
the call to `load_action_queue()` by the `AnimationManager` object.

Now that we have this dictionary of differences that are loaded into
`AnimationManager` via `load_action_queue()`, we need to display them.
First of all, this information needs to be stored since it may not be
used immediately after it is loaded. The dictionary of
differences is assigned to the `differences` attribute. We also make a new
`font` attribute to the `AnimationManager` class since it will need to
display text as well.

Now, we need to actually display these changes in the
`proccess_action()` function. The condition for these changes to be
shown is if the action type is "Attack" or "Heal". Assuming the above
conditions were met, the code loops through the dictionary of
differences drawing the changes in health for each relevant entity (text
is red if the health difference is below 0, green otherwise). The
location for the text is specified in the "Pos" key of the dictionary.

Other Changes:
- Changed the dimensions of all `Knight` class animations from 120x80 to
  90x50
- Fixed a bug with the animations in `Animation`
- Update test cases