Pygame-RPG 30 Refactoring UI
This PR is all about refactoring the current UI system because it became
a tangled mess. This PR aims to make the code more modular and
manageable. It is also designed to be more future proof and easy to
upgrade in the future.

In order to do this, 2 new classes were created `Cursor` and `Submenu`.

`Cursor`: This class controls how the cursor should move in response to
player inputs. It goes through tuple coordinates in a matrix and has
logic for certain special cases.

The main function for the `Cursor` class is the `handle_cursor()`
function. This funciton handles the cursors movement based on the
players inputs. It also has two default flags, `correct` and
`wraparound` both are true by default. If `correct` is true, the cursor
will not move to any "invalid" positions. If `wraparound` is true, the
cursor will "wraparound" the list. For example, if the player presses
right at the last index of a row, it will move the cursor to the first
index of the next row and vice versa.

`Submenu`: This class controls how the submenus in the game operate. It
acts independantly from UIManager which helps in lowering complexity.

The main function for the `Submenu` class is the `handle_cursor()`
function. This function is for handling the cursor interactions in a
submenu which sometimes differ from the norm. This is because a submenu
can have multiple "pages" which can be sifted through by going out of
bounds.