Pygame-RPG 30 Refactoring UI
This PR is all about refactoring the current UI system because it became
a tangled mess. This PR aims to make the code more modular and
manageable. It is also designed to be more future proof and easy to
upgrade in the future.

In order to do this, 2 new classes were created `Cursor` and `Submenu`.

`Cursor`: This class controls how the cursor should move in response to
player inputs. It goes through tuple coordinates in a matrix and has
logic for certain special cases. To achieve this, the `Cursor` class has
the following attributes, `positions`, `rowSlider`, `columnSlider` and
`cursor`.

`positions`: A 2D matrix filled with x, y coordinate tuples that is given
upon initialization.

`rowSlider`: An integer that indicates which "row" in the `positions`
matrix the object is on.

`columnSlider`: An integer that indicates which "column" in the
`positions` matrix the object is on.

`pos`: The tuple that shows the current x, y coordinates of the Cursor
on the screen.

`cursor`: The image that is drawn onto the screen.

The `Cursor` class has the following methods, `reset()`,
`set_new_positions()`, `get_index_from_cursor()` and `handle_cursor()`.

`reset()`: This function sets the `rowSlider` and `columnSlider` values
to 0 and changes the `pos` attribute to ().

`set_new_positions()`: This function takes in a 2D matrix(positions) and
sets `positions` to the given 2D Matrix (assuming the matrix isn't empty
i.e len(positions) > 0). It then resets the `rowSlider` and
`columnSlider` attributes to 0 and sets the `pos` attribute to the first
item in the new matrix.

`get_index_from_cursor()`: This function translates the position of the
cursor from a 2D space to a 1D space by using the `rowSlider` and
`columnSlider` attributes. This function assumes that the "rows" all
have equal length and all indeces before the given index aren't empty
(i.e. there are no holes in our matrix). It returns the result of these
calculations as a number.

`handle_cursor()`: This function takes in a list of player input events
and processes them. This function has 2 default flags `correct` and
`wraparound`.

If `correct` is true then the `handle_cursor()` function will not allow
the `rowSlider` or `columnSlider` to take on invalid values`. For
example if there are only 4 columns, `columnSlider` can never be 4 (we
start at index 0). This behavior is annoying in other contexts so it
can be disabled. Also, the `pos` attribute of the `Cursor` is only ever
changed directly in the function if the `correct` flag is true.

If `wraparound` is true then the `handle_cursor()` function allows for
wraparound to happen. This means that if you reach the last column and
you choose to go right once more, you'll be sent to the first column of
the next row (if it exists). This behavior however does not neccesarily
guranteed a valid index however (due to how the inequalities work).

The `handle_cursor()` function is meant to translate keyboard actions to
changes in `rowSlider` and `columnSlider`. This function returns either
False (normal case), True (if the player pressed enter) and None if the
player pressed escape.

The `Submenu` class has the following attributes, `UI`, `screen`,
`font`, `items`, `page`, `maxPage`, `minIndex`, `maxIndex`,
`constraints` and `submenu`.

`UI`: A string that describes the UI that the submenu spawned from

`screen`: Pygame screen used for drawing

`font`: Pygame font used for writing text on screen

`items`: A list of strings that represent choices in the submenu. This
list can never be empty. In the case that an empty list is passed to the
submenu, it will be replaced with a list with an empty string.

`page`: An integer that keeps track of the "page" the submenu is (this
effectively just indicates the min and max indeces of the `items` list).

`maxPage`: The absolute highest page number that the submenu has.

`minIndex`: The starting index of the `items` list.

`maxIndex`: The ending index of the `items` list. Since Submenus are
designed to be 3x3 matrices, the difference between the min and max
indices is at most 8.

`constraints`: A dictionary of constraints used in drawing the submenu
items as well as indicating the cursor.

`cursor`: `Cursor` object

The `Submenu` class has the following methods, `get_item()`,
`move_to_page()`, `update_page()`, `draw_submenu_items()`, `handle_cursor()` and `draw_UI()`.

`get_item()`: This function uses the values of the `minIndex` and the
`Cursor` to access the specified item in the `items` list.

`move_to_page()`: This function takes in an integer which represents the
page and adjusts the values of the `page`, `minIndex`, `maxIndex` and
sets new positions for the `Cursor` object.

`update_page()`: This function checks if the player has moved to a
different "page". If the player has moved to a different page it
calculates the page and if the page is lower that the `maxPage`
attribute, it calls the `move_to_page()` function with the
calculated page. It also does some math to ensure that the new index for
the player is relative to their previous one. This function returns None
if the calculated page matches the current `page` attribute, `true` if
the `move_to_page()` function was called and `false` if the page was
failed to be changed (i.e. page surpases `maxPage`).

`draw_submenu_items()`: This function uses the items between `minIndex`
and `maxIndex` in the `items` along with some pre defined formulas to
draw the items on the screen using `screen`.

`handle_cursor()`: This function takes in an eventlist. Firstly, this
function copies the state of the `cursor` attribute before calling
`cursor.handle_cursor()` with the inputed eventlist and setting both the
`correct` and `wraparound` flags to false and stores the returned
output. It then calls `update_page()` to see if the page had been
changed and stores the output of the function. If `update_page()`
returned false then the previous state of `cursor` overwrites the
current state (implies that the `cursor` went out of bonds). The
`cursor.handle_cursor()` function is then called with the `correct` boolean set
to true and `wraparound` set to false.

if `cursor.handle_cursor()` returned true, we return the tuple
containing both the `UI`+ (S), `get_item()`. If
`cursor.handle_cursor()` returned None, we return a tuple of `UI`(S),
None`. If `cursor.handle_cursor()` returned false, we return None, None.

`draw_UI()`: This function takes in an eventlist. It then draws the
submenu and calls `handle_cursor()` with the inputted event list. It
then draws the cursor and the submenu items. It processes the choice
made and returns the UI and the choice gotten from `handle_cursor()`.

Other Changes:
- `UIHandler` now has a built in stack for dealing with the UI.
  Essentially, the head of the stack is the one that ACTUALLY processes
  all of the player events. The tail (always UIManager) is passed an
  empty list to process so it can still draw the background stuff.

- `UIHandler` had the interactions code changed

- Algos.py received a new function `construct_matrix_given_parameters()`.
  This function takes in a dictionary of constraints and fills in a
  position matrix for cursors using it.

- Rpg2.py received some changes so that the stack implementation for
  `UIHandler` works.

- `BattleManager` received a new function called
  `get_enemy_pos_matrix()`. This function uses the list of enemies and
  constructs a matrix using it. The constructed matrix is returned.
  There's also a new function called `get_adjusted_knight_pos()`
  which returns the knight's position but adjusted by some constants.

- fixed a glitch in `AnimationManager` where the final frame of the
  animation didn't play causing the positions to be offset. This broke
  how the cursor worked and made it possible to index out of bounds.