Pygame-RPG 30 Refactoring UI
This PR is all about refactoring the current UI system because it became
a tangled mess. This PR aims to make the code more modular and
manageable. It is also designed to be more future proof and easy to
upgrade in the future.

In order to do this, 2 new classes were created `Cursor` and `Submenu`.

`Cursor`: This class controls how the cursor should move in response to
player inputs. It goes through tuple coordinates in a matrix and has
logic to control how things behave.

`Submenu`: This class controls how the submenus in the game operate. It
acts independantly from UIManager which helps in lowering complexity.